# Conflicts:
#	app/src/main/res/values-fr/strings.xml

+++ app/src/main/java/fr/upem/crazygame/provider/DatabaseHelper.java
@@ -4,6 +4,9 @@ import android.content.ContentValues;
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
+import android.util.Log;
+
+import static fr.upem.crazygame.provider.ProviderDataGame.CONTENT_PROVIDER_DB_VERSION;
 
 
 /**
@@ -12,11 +15,12 @@ import android.database.sqlite.SQLiteOpenHelper;
 public class DatabaseHelper extends SQLiteOpenHelper {
 
         DatabaseHelper(Context context) {
-            super(context, ProviderDataGame.CONTENT_PROVIDER_DB_NAME, null, ProviderDataGame.CONTENT_PROVIDER_DB_VERSION);
+            super(context, ProviderDataGame.CONTENT_PROVIDER_DB_NAME, null, CONTENT_PROVIDER_DB_VERSION);
         }
 
         @Override
         public void onCreate(SQLiteDatabase db) {
+            Log.d("test", "PROVIDERRRRRRRRRRR");
             db.execSQL("CREATE TABLE " + ProviderDataGame.CONTENT_PROVIDER_TABLE_NAME
                     + " (" + GameCrazyGameColumns.NAME_GAME + ","
                     + GameCrazyGameColumns.GAME + " INTEGER,"
@@ -40,12 +44,22 @@ public class DatabaseHelper extends SQLiteOpenHelper {
             contentValuesMixWord.put(GameCrazyGameColumns.GAME_LAST_PLAY, 0);
 
             db.insert(ProviderDataGame.CONTENT_PROVIDER_TABLE_NAME, null, contentValuesMixWord);
+
+            ContentValues contentValuesShakeGame = new ContentValues();
+            contentValuesShakeGame.put(GameCrazyGameColumns.NAME_GAME, GameCrazyGameColumns.NAME_SHAKEGAME);
+            contentValuesShakeGame.put(GameCrazyGameColumns.GAME, 0);
+            contentValuesShakeGame.put(GameCrazyGameColumns.GAME_WIN, 0);
+            contentValuesShakeGame.put(GameCrazyGameColumns.GAME_LAST_PLAY, 0);
+
+            db.insert(ProviderDataGame.CONTENT_PROVIDER_TABLE_NAME, null, contentValuesShakeGame);
         }
 
         // Cette méthode sert à gérer la montée de version de notre base
         @Override
         public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-            db.execSQL("DROP TABLE IF EXISTS " + ProviderDataGame.CONTENT_PROVIDER_TABLE_NAME);
+            if (oldVersion < newVersion) {
+                db.execSQL("DROP TABLE IF EXISTS " + ProviderDataGame.CONTENT_PROVIDER_TABLE_NAME);
+            }
             onCreate(db);
         }
     }
